requirements.txt
telethon==1.36.0
python-telegram-bot==21.5
discord.py-self==2.0.1
aiohttp==3.10.5
python-dotenv==1.0.1
asyncio==3.4.3
